requirements.txt
amqp==2.3.2
appdirs==1.4.0
billiard==3.5.0.2
celery==4.0.2
certifi==2017.7.27.1
chardet==3.0.4
Django==2.0.1
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-extra-views==0.9.0
django-impersonate==1.3
django-simple-history==2.3.0
idna==2.5
kombu==4.0.2
packaging==16.8
pytz==2017.2
redis==2.10.5
requests==2.18.3
six==1.10.0
urllib3==1.22
vine==1.1.3
django-debug-toolbar==1.9.1
